requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.5
pydantic-settings==2.7.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httpx==0.28.1
python-multipart==0.0.19
python-dotenv==1.0.1
cryptography==45.1.0
websockets==14.1
youtube-comment-downloader==0.1.74
supabase==2.9.1
redis==5.2.1
pytest==8.3.4
pytest-asyncio==0.25.0